+++ src/commands/mod.rs
@@ -1,4 +1,6 @@
 pub mod chart;
+pub mod currency;
 pub mod price;
 pub mod price_all;
+pub mod send_all;
 pub mod start;