+++ src/commands/price_all.rs
@@ -1,15 +1,22 @@
 use crate::models::user::User;
+use log::{debug, info};
 use reqwest::Url;
 use std::env;
 
+/// /priceall command handler
+/// send info about all user currency
 pub async fn price_all_command(user: User) -> String {
+    info!("price_all_command");
     if user.currency.is_empty() {
         return "You don't have any currency, type /addcurency curency-name".to_string();
     }
     let result = get_currency_price_multi(user.currency).await;
     match result {
         Ok(res) => res,
-        Err(e) => e.to_string(),
+        Err(e) => {
+            debug!("price all error {}", e);
+            "Error, maybe you don't have any valid currency".to_string()
+        }
     }
 }