+++ src/db.rs
@@ -1,4 +1,6 @@
 use crate::models::user::User;
+use futures::stream::StreamExt;
+use log::debug;
 use mongodb::bson::Document;
 use mongodb::{
     bson::{doc, Bson},
@@ -44,16 +46,41 @@ impl DatabaseManager {
             "updated_at": user.updated_at,
         };
 
-        collection.insert_one(user_doc, None).await?;
+        let res = collection.insert_one(user_doc, None).await;
         // info!("Inserted user: {:?}", res.inserted_id);
-        // info!("Inserted user: {:?}", res);
+
+        match res {
+            Ok(_) => (),
+            Err(err) => debug!("user already exists: {}", err),
+        }
+
         Ok(())
     }
 
-    pub async fn get_user(&self, user_id: i64) -> Result<User, Box<dyn Error>> {
+    pub async fn get_user(&self, user_id: i64) -> Option<User> {
         let collection: Collection<User> = self.db.collection("user");
-        let user = collection.find_one(doc! {"user_id": user_id}, None).await?;
-        Ok(user.unwrap())
+        let user_query = collection.find_one(doc! {"user_id": user_id}, None).await;
+
+        match user_query {
+            Ok(Some(user)) => {
+                // debug!("user Ok: {:?}", user);
+                Some(user)
+            }
+            Ok(None) => {
+                let usr = User::new(user_id, "".to_string(), vec![]);
+                match usr.save(self.clone()).await {
+                    Ok(_) => Some(usr),
+                    Err(err) => {
+                        debug!("user save error: {:?}", err);
+                        None
+                    }
+                }
+            }
+            Err(err) => {
+                debug!("user save error2: {:?}", err);
+                None
+            }
+        }
     }
 
     // change user currency
@@ -63,14 +90,24 @@ impl DatabaseManager {
         currency: String,
     ) -> Result<(), Box<dyn Error>> {
         let collection: Collection<User> = self.db.collection("user");
-        let mut user = self.get_user(user_id).await?;
+        let user_query = self.get_user(user_id).await;
+        if user_query.is_none() {
+            return Err("Problem with user saving".into());
+        }
+        let mut user = user_query.unwrap();
         user.currency.push(currency);
-        user.currency.dedup();
+        let mut unique_currency: Vec<String> = Vec::new();
+        for c in user.currency.into_iter() {
+            if !unique_currency.contains(&c) {
+                unique_currency.push(c);
+            }
+        }
+        user.currency = unique_currency;
         let user_doc = self.update_user_doc(user).await;
-        //collection.update_one(doc! {"user_id": user_id}, user_doc, None).await?;
         collection
             .update_one(doc! {"user_id": user_id}, doc! {"$set": user_doc}, None)
             .await?;
+        //debug!("Updated user: {:?}", res.modified_count);
         Ok(())
     }
 
@@ -80,7 +117,12 @@ impl DatabaseManager {
         currency_to_remove: String,
     ) -> Result<(), Box<dyn Error>> {
         let collection: Collection<User> = self.db.collection("user");
-        let mut user = self.get_user(user_id).await?;
+        let user_query = self.get_user(user_id).await;
+
+        if user_query.is_none() {
+            return Err("Problem with user saving".into());
+        }
+        let mut user = user_query.unwrap();
 
         // Отфильтровать список валют, удаляя указанную валюту
         user.currency
@@ -104,5 +146,20 @@ impl DatabaseManager {
             "created_at": user.created_at,
             "updated_at": mongodb::bson::DateTime::now(),
         }
+    }
+
+    pub async fn get_all_users(
+        &self,
+        filter: Option<Document>,
+    ) -> Result<Vec<User>, Box<dyn Error>> {
+        let collection: Collection<User> = self.db.collection("user");
+        let mut cursor = collection.find(filter, None).await?;
+        let mut users_vec: Vec<User> = Vec::new();
+        while let Some(result) = cursor.next().await {
+            let user = result?;
+            users_vec.push(user);
+        }
+
+        Ok(users_vec)
     }
 }