+++ src/models/user.rs
@@ -1,5 +1,8 @@
+use crate::db::DatabaseManager;
+use log::debug;
 use mongodb::bson;
 use serde::{Deserialize, Serialize};
+use std::error::Error;
 
 /// User model
 ///
@@ -16,12 +19,17 @@ use serde::{Deserialize, Serialize};
 /// * `new` - Create new user
 ///
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct User {
+    /// User id
     pub user_id: i64,
+    /// User username
     pub username: String,
+    /// User currency
     pub currency: Vec<String>,
+    /// User created at
     pub created_at: bson::DateTime,
+    /// User updated at
     pub updated_at: bson::DateTime,
 }
 
@@ -43,5 +51,14 @@ impl User {
             created_at: bson::DateTime::now(),
             updated_at: bson::DateTime::now(),
         }
+    }
+
+    pub async fn save(&self, db: DatabaseManager) -> Result<(), Box<dyn Error>> {
+        let res = db.insert_user(self.clone()).await;
+        match res {
+            Ok(_) => (),
+            Err(_) => debug!("user already exists"),
+        }
+        Ok(())
     }
 }