+++ src/tests/mod.rs
@@ -1,2 +1,3 @@
 #[cfg(test)]
 pub mod currency_tests;
+pub mod twitter_tests;